+++ packages/shield-controller/src/backend.test.ts
@@ -50,7 +50,7 @@ function setup({
     getCoverageResultTimeout,
     getCoverageResultPollInterval,
     fetch,
-    baseUrl: 'https://rule-engine.metamask.io',
+    baseUrl: 'https://ruleset-engine.api.cx.metamask.io',
     captureException: mockCaptureException,
   });