+++ packages/wallet-cli/src/daemon/wallet-factory.test.ts
@@ -121,6 +121,12 @@ describe('createWallet', () => {
     );
     expect(instanceOptions.transactionController?.disableSwaps).toBe(true);
     expect(instanceOptions.transactionController?.hooks).toStrictEqual({});
+    expect(instanceOptions.shieldController.baseUrl).toBe(
+      'https://ruleset-engine.api.cx.metamask.io',
+    );
+    expect(instanceOptions.shieldController.fetchFunction).toBe(
+      globalThis.fetch,
+    );
     expect(ClientConfigApiService).toHaveBeenCalled();
 
     await dispose();